+++ src/core/components/link/README.md
@@ -8,61 +8,15 @@
 $ yarn add @guardian/src-link
 ```
 
-## Use
+## Usage
 
-```js
-import { Link, ButtonLink } from '@guardian/src-link';
-import { SvgArrowRightStraight } from '@guardian/src-icons';
+### API
 
-const Navigation = () => (
-    <>
-        <Link
-            priority="primary"
-            subdued={true}
-            icon={<SvgArrowRightStraight />}
-            iconSide="right"
-            href="/read-more"
-        >
-            Click me
-        </Link>
-        <ButtonLink
-            priority="primary"
-            subdued={true}
-            icon={<SvgArrowRightStraight />}
-            iconSide="right"
-            onClick={() => console.log('Clicked')}
-        >
-            Click me
-        </ButtonLink>
-    </>
-);
-```
+See [storybook](https://guardian.github.io/source/?path=/docs/source-src-link-link--demo)
 
-## `Link` and `ButtonLink` Props
+### How to use
 
-### `priority`
-
-**`"primary" | "secondary"`** _= "primary"_
-
-Informs users of how important a link is
-
-### `subdued`
-
-**`boolean`** _= false_
-
-Whether link is subdued (no underline)
-
-### `icon`
-
-**`JSX.Element`**
-
-An icon that appears in the link, alongside text
-
-### `iconSide`
-
-**`"left" | "right"`** _= "left"_
-
-The side of the link on which the icon appears
+For context and visual guides relating to usage see the [Source Design System website](https://theguardian.design/2a1e5182b/p/43c26b-link/b/048fd1).
 
 ## Supported themes
