--- src/core/components/link/button-link.stories.tsx
@@ -1,7 +1,0 @@
-export * from './stories/inline-button-link';
-export * from './stories/priority-button-links';
-export * from './stories/text-and-icon-button-links';
-
-export default {
-	title: 'ButtonLink',
-};
