+++ src/core/components/link/tsconfig.json
@@ -6,7 +6,7 @@
     "emitDeclarationOnly": true
   },
   "include": [".", "../../foundations/types/themes.d.ts"],
-  "exclude": ["dist", "stories", "stories.tsx"],
+  "exclude": ["dist", "*.stories.tsx"],
   "references": [
     {
       "path": "../../helpers"
